Cleaning of UniformMass [This commit may break a lot of things]
The existing implementation of Uniform is a real mess.
Among obvious problem:
	- several totally duplicated implementation (eg: loading the mass matrix from a file, getMomentum, draw)
	- no warning nor messages indicating the filename is only valid for Rigid3fd only templates.
	- some functions (eg getMomentum) are defined but only correctly implemented for certain templates all
          the other template just return 0,0,0 without any warning nor error message.
	- implicit handling when both mass and totalMass are set (to know what happens user has to read the source code).

So I tried this version to factorize that a bit.

The most hard things was to be able to handle the duplicated implementations of getMomentum
to handle in an unified way the dimensions of the template type. The current implementation
is not a really satisfactory hack but I made it to avoid touching too much to RigidTypes.h

The underlying problem is that Rigid<2,> and Rigid<3,> are specialized template with different
public API which make that a single template cannot handle Rigid<3,...> and Rigid<2,...>.

The trick I used consist in using std::enable_if from C++11. When the template
is "compatible with the implementation" std::enable_if.